Last Commit:
Fixed sounds having doppler effect, and most of the last bugs. Known bugs still remaning:

If dropping an object while holding it, one has to leave the colliderbox to pick it back up again.
If you drop two buildings simulataniously, they might not removing the component buildings if you make a combo out of them.
Putting a park next to a luxury house might sometimes turn the luxury house into a sculpture.
Soundeffects on the building pallet may sometimes have a very low sound.
If the player holds the start block while starting the game, the text remains on the controller, and the controller might be unable to pick up other tiles or objects.